+++ runners/runner_asset_allocation.py
@@ -1,14 +1,12 @@
 import pandas as pd
+import numpy as np
 import quantkit.loader.runner as loader
 import quantkit.utils.logging as logging
-import quantkit.data_sources.snowflake as snowflake
-import quantkit.finance.data_sources.quandl_datasource.quandl_datasource as quds
 import quantkit.utils.mapping_configs as mapping_configs
 import quantkit.asset_allocation.strategies.momentum as momentum
 import quantkit.asset_allocation.strategies.pick_all as pick_all
 import quantkit.utils.mapping_configs as mapping_configs
-import quantkit.utils.snowflake_utils as snowflake_utils
-import numpy as np
+import quantkit.asset_allocation.universe.universe as universe_datasource
 
 
 class Runner(loader.Runner):
@@ -25,17 +23,19 @@ class Runner(loader.Runner):
         """
         super().init(local_configs)
 
-        # connect quandl datasource
-        self.universe = dict()
-
         self.annualize_factor = mapping_configs.annualize_factor_d.get(
-            self.params["quandl_datasource_prices"]["frequency"], 252
+            self.params["prices_datasource"]["frequency"], 252
         )
         self.rebalance_window = mapping_configs.rebalance_window_d.get(
-            self.params["rebalance"]
+            self.params["prices_datasource"]["rebalance"]
         )
 
         self.strategies = dict()
+
+        # connect portfolio datasource
+        self.portfolio_datasource = universe_datasource.Universe(
+            params=self.params["universe_datasource"], api_settings=self.api_settings
+        )
 
         # iterate over dataframes and create objects
         logging.log("Start Iterating")
@@ -45,10 +45,11 @@ class Runner(loader.Runner):
         """
         iterate over DataFrames and create connected objects
         """
+        self.iter_parent_issuers()
         self.iter_portfolios()
-        self.create_universe()
         self.iter_msci()
-        self.iter_quandl()
+        self.iter_prices()
+        self.iter_fundamentals()
         self.iter_holdings()
         self.iter_securities()
         self.iter_cash()
@@ -71,13 +72,11 @@ class Runner(loader.Runner):
             - allocation_models: list
                 list of all weighting strategies
         """
-        for strategy in self.params["strategies"]:
-            strat_params = self.params["strategies"][strategy]
-            strat_params["frequency"] = self.params["quandl_datasource_prices"][
-                "frequency"
-            ]
+        for strategy, strat_params in self.params["strategies"].items():
+            strat_params["frequency"] = self.params["prices_datasource"]["frequency"]
+            strat_params["rebalance"] = self.params["prices_datasource"]["rebalance"]
             strat_params["universe"] = self.portfolio_datasource.all_tickers
-            strat_params["rebalance_dates"] = self.rebalance_dates
+            strat_params["rebalance_dates"] = self.prices_datasource.rebalance_dates
             strat_params["trans_cost"] = self.params["trans_cost"]
             strat_params["weight_constraint"] = self.params[
                 "default_weights_constraint"
@@ -87,107 +86,32 @@ class Runner(loader.Runner):
             elif strat_params["type"] == "pick_all":
                 self.strategies[strategy] = pick_all.PickAll(strat_params)
 
-    def iter_holdings(self) -> None:
-        super().iter_holdings()
-        # filter for securities that match msci ticker
-        for s, sec_store in self.portfolio_datasource.securities.items():
-            ticker = sec_store.information["Ticker Cd"]
-            if ticker in self.portfolio_datasource.all_tickers:
-                self.universe[ticker] = sec_store
-
-    def iter_quandl(self) -> None:
-        """
-        - iterate over quandl data
-            - attach quandl information to company in self.quandl_information
-        - create price DataFrame
-        - create return DataFrame with sorted universe in columns
-        """
-        super().iter_quandl()
-
-        # price data
-        self.price_data = self.quandl_datasource_prices.df.pivot(
-            index="date", columns="ticker", values="closeadj"
-        )
-
-        # return data
-        self.return_data = self.price_data.pct_change(1)
-
-        # filter universe for companies that have quandl data
-        self.portfolio_datasource.all_tickers = list(
-            set(self.return_data.columns)
-            & set(self.quandl_datasource.df["ticker"])
-            & set(self.portfolio_datasource.all_tickers)
-        )
-
-        # order return data
-        self.return_data = self.return_data[self.portfolio_datasource.all_tickers]
-
-        # rebalance dates -> last trading day of month
-        self.rebalance_dates = list(
-            self.return_data.groupby(
-                pd.Grouper(
-                    freq=mapping_configs.pandas_translation[self.params["rebalance"]]
-                )
-            )
-            .tail(1)
-            .index
-        )
-
-        # fundamental dates -> date + 3 months
-        self.fundamental_dates = list(
-            self.quandl_datasource.df["release_date"].sort_values().unique()
-        )
-        self.next_fundamental_date = 0
-
-        # initialize market caps
-        self.market_caps = np.ones(shape=len(self.portfolio_datasource.all_tickers))
-
-    def create_universe(self) -> None:
-        """
-        - Load Portfolio Data from snowflake
-        - create universe based on indexes from params file
-        """
-        if self.params["sustainable_universe"]:
-            df = snowflake_utils.load_from_snowflake(
-                database="SANDBOX_ESG",
-                schema="ESG_SCORES_THEMES",
-                table_name="Sustainability_Framework_Detailed",
-                local_configs=self.local_configs,
-            )
-            # all tickers in index which are labeled green or blue
-            self.portfolio_datasource.all_tickers = list(
-                df[
-                    (df["Portfolio ISIN"].isin(self.params["universe"]))
-                    & (df["SCLASS_Level2"].isin(["Transition", "Sustainable Theme"]))
-                ]["Ticker"].unique()
-            )
-
     def run_strategies(self) -> None:
         """
         Run and backtest all strategies
         """
-        for date, row in self.return_data.iterrows():
+        for date, row in self.prices_datasource.return_data.iterrows():
             r_array = np.array(row)
-
-            # assign new market weights each quarter
-            if date >= self.fundamental_dates[self.next_fundamental_date]:
-                df = self.quandl_datasource.df.pivot(
-                    index="release_date", columns="ticker", values="marketcap"
-                )
-                df = df.loc[self.fundamental_dates[self.next_fundamental_date]]
-
-                df = df[self.portfolio_datasource.all_tickers]
-                self.market_caps = np.array(df)
-                self.next_fundamental_date += 1
+            market_caps = self.fundamentals_datasource.outgoing_row(date)
+            index_components = self.portfolio_datasource.outgoing_row(date)
 
             # assign returns to strategies and backtest
             for strat, strat_obj in self.strategies.items():
-                strat_obj.assign(date, r_array)
-                strat_obj.backtest(
-                    date,
-                    self.market_caps,
-                    mapping_configs.annualize_factor_d[self.params["rebalance"]],
+                strat_obj.assign(date, r_array, index_components)
+                strat_obj.backtest(date, market_caps)
+
+        # assign weights to security objects
+        for strat, strat_obj in self.strategies.items():
+            for allo, allo_obj in strat_obj.allocation_engines_d.items():
+                allocation_df = pd.DataFrame.from_dict(
+                    allo_obj.allocations_history,
+                    orient="index",
+                    columns=self.portfolio_datasource.all_tickers,
                 )
+                for sec in allocation_df.columns:
+                    self.portfolio_datasource.tickers[sec].allocation_df[
+                        f"{strat}_{allo}"
+                    ] = allocation_df[sec]
 
     def run(self) -> None:
         """