+++ utils/configs.py
@@ -3,17 +3,18 @@ import os
 import datetime
 from pandas.tseries.offsets import BDay
 import quantkit.utils.util_functions as util_functions
+from typing import Union
 
 
-def read_configs(local_configs: str = "") -> dict:
+def read_configs(local_configs: Union[str, dict] = "") -> dict:
     """
     read configs file from quantkit\\utils\\configs.json
     if there are local changes, overwrite existing configs file
 
     Parameters
     ----------
-    local_configs: str, optional
-        path to a local configarations file
+    local_configs: str | dict, optional
+        path to a local configarations file or local configs as dict
 
     Returns
     -------
@@ -24,7 +25,9 @@ def read_configs(local_configs: str = "") -> dict:
         configs = json.load(f_in)
 
     # local configs file exists
-    if os.path.isfile(local_configs):
+    if isinstance(local_configs, dict):
+        configs = util_functions.replace_dictionary(local_configs, configs)
+    elif os.path.isfile(local_configs):
         with open(local_configs) as f_in:
             configs_local = json.load(f_in)