+++ CHANGELOG.md
@@ -11,6 +11,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Removed
 
+## [1.4.0] - 2023-11-24
+### Added
+- JPM CEMBI BROAD DIVERSE and PM EMBI GLOBAL DIVERSIFI added to portfolio df
+- Universe Datasource in asset allocation tool
+- Allow custom universe
+- Make universe change over time 
+- Allocation history per strategy attached to each security
+### Fixed
+### Changed
+- Split QuandlDataSource into PricesDataSource and FundamentalsDataSource
+- Use numpy array for date iterations
+### Removed
+
 ## [1.3.3] 2023-11-16
 ### Added
 - Snowflake util function to update daily DataFrame
@@ -178,6 +191,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 [unreleased]: https://gitlab.com/tcw-group/quant-research/quantkit/-/compare/main...develop?from_project_id=46798372&straight=false
+[1.4.0]: https://gitlab.com/tcw-group/quant-research/quantkit/-/releases/1.4.0
 [1.3.3]: https://gitlab.com/tcw-group/quant-research/quantkit/-/releases/1.3.3
 [1.3.2]: https://gitlab.com/tcw-group/quant-research/quantkit/-/releases/1.3.2
 [1.3.1]: https://gitlab.com/tcw-group/quant-research/quantkit/-/releases/1.3.1